+++ core-ui/src/main/java/com/agileburo/anytype/core_ui/features/page/styling/StylingEvent.kt
@@ -13,8 +13,9 @@ sealed class StylingEvent {
     sealed class Markup : StylingEvent() {
         object Bold : Markup()
         object Italic : Markup()
-        object Strikethrough : Markup()
+        object StrikeThrough : Markup()
         object Code : Markup()
+        object Link : Markup()
     }
 
     sealed class Coloring : StylingEvent() {