+++ core-ui/src/test/java/com/agileburo/anytype/core_ui/HighlightingBlockTest.kt
@@ -105,6 +105,7 @@ class HighlightingBlockTest {
             onFooterClicked = {},
             onTextInputClicked = {},
             onPageIconClicked = {},
+            onProfileIconClicked = {},
             onTogglePlaceholderClicked = {},
             onToggleClicked = {},
             onParagraphTextChanged = { _, _ -> },